requirements.txt
asgiref==3.2.9
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
django-markdownx==3.0.1
flake8==3.8.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mypy==0.780
mypy-extensions==0.4.3
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
whitenoise==5.1.0
wrapt==1.12.1
